Moved subscription specific compoents to the /app/upgrade folder
